댓글 삭제 완료

+++ src/main/java/com/todoapp/todoapp/service/AppCommentService.java
@@ -1,5 +1,6 @@
 package com.todoapp.todoapp.service;
 
+import com.todoapp.todoapp.customException.uqualsException;
 import com.todoapp.todoapp.dto.comment.CommentRequestDto;
 import com.todoapp.todoapp.dto.comment.CommentResponseDto;
 import com.todoapp.todoapp.entity.Card;
@@ -37,6 +38,7 @@ public class AppCommentService {
         comments.update(requestDto);
 
 //        이렇게 댓글을 수정후 체크해보는데 카드입장에서는 바로 즉시 수정한게 반영되는데 유저입장도 수정된게 반영되긴하나 카드처럼 즉시는 안됩니다.
+        //지금 테스트는 한게시글에 대해서 동일 유저가 댓글 여러개 적었을떄.
 //         (예시 출력결과 ↓) 영속성컨텍스트랑 관련 있는것같긴한데 정확히 무슨 이유인지 모르겠습니다...
 //        카드입장체크 Good다시수정2
 //        카드입장체크 Hi수정
@@ -55,12 +57,16 @@ public class AppCommentService {
         return new CommentResponseDto(comments);
     }
 
-    public void deleteComment(Long id) {
+    public void deleteComment(Long id, User user) throws uqualsException {
         Comments comments = findComment(id);
+
+        if(!comments.getUser().getUsername().equals(user.getUsername())){
+            throw new uqualsException();
+        }
+
         commentRepository.delete(comments);
+
     }
-
-
 
     private Comments findComment(Long id) {
         return commentRepository.findById(id).orElseThrow(() ->
@@ -73,5 +79,6 @@ public class AppCommentService {
                 new IllegalArgumentException("해당 앱카드는 없습니다")
         );
     }
+
 
 }
